pyproject.toml
[tool.poetry]
name = "zurihngxtask1"
version = "0.1.0"
description = ""
authors = ["Ayomide <ayomidet905@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
django-rest-framework = "^0.1.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
